+++ constrainedterm.cpp
@@ -52,7 +52,7 @@ void ConstrainedTerm::smtNarrowSearchHelper(RewriteSystem &rs,
   }
   if (depth < maxDepth) {
     vector<ConstrainedSolution> sols = this->smtNarrowSearch(rs);
-    for (int i = 0; i < sols.size(); ++i) {
+    for (int i = 0; i < (int)sols.size(); ++i) {
       ConstrainedTerm(sols[i].term, sols[i].getFullConstraint(*this)).smtNarrowSearchHelper(rs, minDepth, maxDepth, depth + 1, result);
     }
   }
@@ -67,7 +67,7 @@ void ConstrainedTerm::smtNarrowSearchHelper(CRewriteSystem &crs,
   } 
   if (depth < maxDepth) {
     vector<ConstrainedSolution> sols = this->smtNarrowSearch(crs);
-    for (int i = 0; i < sols.size(); ++i) {
+    for (int i = 0; i < (int)sols.size(); ++i) {
       ConstrainedTerm(sols[i].term, sols[i].getFullConstraint(*this)).smtNarrowSearchHelper(crs, minDepth, maxDepth, depth + 1, result);
     }
   }
